+++ src/assets.rs
@@ -1,0 +1,9 @@
+pub mod lua {
+    pub mod input {
+        pub const EXAMPLE_EN: &str = include_str!("../assets/lua/input/example_en.lua");
+        pub const EXAMPLE_ZH_HANS: &str = include_str!("../assets/lua/input/example_zh-hans.lua");
+    }
+    pub mod lib {
+        pub const TALK: &str = include_str!("../assets/lua/lib/talk.lua");
+    }
+}