+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example variants in languages.
 - `path` field for the Lua class `Image`.
 - `"ko"` and `"zh-Hant"` to `lang` field.
+- Add `luatalk::assets` for asset strings.
 
 ### Changed